- implementation of the quest system
NOTE: now it's possible to get a quest using either the command "@quest [questId]" or using the mission receiver item, witch you can buy it from merchant or just drop it (if you drop it don't forget to set the extraStats value). it will only be available the quests loaded in the Static Database.
IT'S NOT POSSIBLE TO DO QUESTS YET! (this still needs to be implemented)

git-svn-id: http://reunion.googlecode.com/svn/trunk@259 d5b2e400-7aba-11de-9499-4f8beaa274e3